Stabilized and kind of improved wakeup process.
* Fixed relatively minor issues from Android Studio's "Inspect Code"

* In app/build.gradle.kts:

    - Updated dependencies to use "-ktx"

* In activealarm/NacWakeupProcess.kt:

    - Added "restrictVolumeHandler" member variable since unable to detect when volume change events occur.  Instead, will have a handler that runs every second and if the volume does not match what it should be (louder volume is also acceptable), then will force the volume back to what it should be.

    - Added "hasGraduallyIncreaseVolumeStarted" member variable to check if the gradual volume increase process has started.  Since "restrictVolumeHandler" will run before "graduallyIncreaseVolumeHandler", need to have this check so that it knows not to restrict volume yet.

    - Moved NacTextToSpeech listener to the class instantiation.

    - In cleanup(), added audio attributes call to revertVolume() as well as cleaning up the new handler.

    - In graduallyIncreaseVolume(), changed logic so that it gets the current volume from the stream (this is normal), but if the volume should be restricted, then it gets it from the "volumeToRestrictChangeTo" member variable.  This is so that the volume cannot be continuously decreased forever.

      Added check to see if the current volume is less than the alarm volume, and if so, increase the volume.

      Always call the gradually increase handler so that these checks are constantly occurring at the cadence.

    - Added restrictVolume() and setupRestrictVolume().

    - Added setupVolume() to save the current volume, then set the restrict volume, then setup the gradually increase volume and restrict volume (if those options are set in the alarm).

    - Renamed start() to simpleStart() and renamed start(NacAlarm) to start(), since the alarm is passed into the constructor.

* In main/NacMainActivity.kt:

    - Fixed issue where in onCreateContextMenu(), "menuInfo" was not nullable.

* In media/NacAudioAttributes.kt:

    - Reworked code so that things make more sense and are organized better.  Got rid of things that didn't make sense like have the volume level negative and have a "canVolumeChange" member variable that checks if the volume level is greater than or equal to 0.

    - Added saveCurrentVolume()

    - Renamed toStreamVolume() to alarmToStreamVolume()

    - Renamed setVolume() to setStreamVolume().

+++ app/src/main/java/com/nfcalarmclock/shared/NacSharedPreferences.kt
@@ -93,16 +93,16 @@ class NacSharedPreferences(
 	/**
 	 * @see .getAutoDismissTime
 	 */
-	val autoDismissTime: Int
+	val autoDismissTime: Long
 		get()
 		{
 			return if (autoDismissIndex < 5)
 			{
-				autoDismissIndex
+				autoDismissIndex.toLong()
 			}
 			else
 			{
-				(autoDismissIndex - 4) * 5
+				(autoDismissIndex - 4) * 5L
 			}
 		}
